Fix jmh benchmarks
